+++ examples/showoci/showoci_service.py
@@ -224,6 +224,7 @@ class ShowOCIService(object):
     # database
     C_DATABASE = "database"
     C_DATABASE_DBSYSTEMS = "dbsystems"
+    C_DATABASE_EXADATA = "exadata"
     C_DATABASE_AUTONOMOUS = "autonomous"
     C_DATABASE_NOSQL = "nosql"
     C_DATABASE_MYSQL = "mysql"
@@ -333,6 +334,7 @@ class ShowOCIService(object):
         {'shape': 'Exadata.Full3.400', 'cpu': 400, 'memory': 5760, 'storage': 598},
         {'shape': 'Exadata.Half3.200', 'cpu': 200, 'memory': 2880, 'storage': 298},
         {'shape': 'Exadata.Quarter3.100', 'cpu': 100, 'memory': 1440, 'storage': 149},
+        {'shape': 'Exadata.X8M', 'cpu': 100, 'memory': 1440, 'storage': 149},
         {'shape': 'Exadata.Base.48', 'cpu': 48, 'memory': 720, 'storage': 74.8},
         {'shape': 'VM.CPU3.1', 'cpu': 6, 'memory': 90, 'storage': 0},
         {'shape': 'VM.CPU3.2', 'cpu': 12, 'memory': 180, 'storage': 0},
@@ -624,7 +626,7 @@ class ShowOCIService(object):
                     print("Error, OCI SDK version " + self.oci_compatible_version + " required !")
                     print("OCI SDK Version installed = " + self.get_oci_version())
                     print("Please use below command to upgrade OCI SDK:")
-                    print("   pip install --upgrade oci")
+                    print("   python -m pip install --upgrade oci")
                     print("")
                     print("Aboting.")
                     print("*********************************************************************")
@@ -2666,7 +2668,8 @@ class ShowOCIService(object):
                     try:
                         arrsecs = oci.pagination.list_call_get_all_results(
                             virtual_network.list_network_security_group_security_rules,
-                            arr.id
+                            arr.id,
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                         ).data
 
                     except oci.exceptions.ServiceError as e:
@@ -2918,7 +2921,8 @@ class ShowOCIService(object):
                 try:
                     arrs = oci.pagination.list_call_get_all_results(
                         virtual_network.list_drg_attachments,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -3445,12 +3449,14 @@ class ShowOCIService(object):
                     arrs = oci.pagination.list_call_get_all_results(
                         compute.list_instances,
                         compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                     consoles = oci.pagination.list_call_get_all_results(
                         compute.list_instance_console_connections,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -3585,7 +3591,8 @@ class ShowOCIService(object):
                         compute.list_images,
                         compartment['id'],
                         sort_by="DISPLAYNAME",
-                        lifecycle_state=oci.core.models.Image.LIFECYCLE_STATE_AVAILABLE
+                        lifecycle_state=oci.core.models.Image.LIFECYCLE_STATE_AVAILABLE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -3610,7 +3617,7 @@ class ShowOCIService(object):
                            'region_name': str(self.config['region']),
                            'defined_tags': [] if arr.defined_tags is None else arr.defined_tags,
                            'freeform_tags': [] if arr.freeform_tags is None else arr.freeform_tags,
-                           'base_image_name': (str(compute.get_image(arr.base_image_id).data.display_name) if arr.base_image_id else "")
+                           'base_image_name': (str(compute.get_image(arr.base_image_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data.display_name) if arr.base_image_id else "")
                            }
                     data.append(val)
                     cnt += 1
@@ -3651,7 +3658,11 @@ class ShowOCIService(object):
                 autos = []
                 try:
                     # pagination didn't work on auto scaling code
-                    autos = oci.pagination.list_call_get_all_results(autoscaling.list_auto_scaling_configurations, compartment['id']).data
+                    autos = oci.pagination.list_call_get_all_results(
+                        autoscaling.list_auto_scaling_configurations,
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                    ).data
 
                 except oci.exceptions.ServiceError as e:
                     if self.__check_service_error(e.code):
@@ -3781,7 +3792,8 @@ class ShowOCIService(object):
                 try:
                     configs = oci.pagination.list_call_get_all_results(
                         compute_manage.list_instance_configurations,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 # inst pool and inst config service often goes down, not marking warning
@@ -3896,7 +3908,8 @@ class ShowOCIService(object):
                 try:
                     pools = oci.pagination.list_call_get_all_results(
                         compute_manage.list_instance_pools,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 # inst pool and inst config service often goes down, not marking warning
@@ -3965,7 +3978,8 @@ class ShowOCIService(object):
                         arrs = oci.pagination.list_call_get_all_results(
                             compute.list_boot_volume_attachments,
                             ad['name'],
-                            compartment['id']
+                            compartment['id'],
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                         ).data
 
                     except oci.exceptions.ServiceError as e:
@@ -4017,7 +4031,8 @@ class ShowOCIService(object):
                 try:
                     arrs = oci.pagination.list_call_get_all_results(
                         compute.list_volume_attachments,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -4152,8 +4167,11 @@ class ShowOCIService(object):
 
                 arrs = []
                 try:
-                    arrs = oci.pagination.list_call_get_all_results(compute.list_vnic_attachments,
-                                                                    compartment['id']).data
+                    arrs = oci.pagination.list_call_get_all_results(
+                        compute.list_vnic_attachments,
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                    ).data
                 except oci.exceptions.ServiceError as e:
                     if self.__check_service_error(e.code):
                         self.__load_print_auth_warning()
@@ -4245,7 +4263,8 @@ class ShowOCIService(object):
                         boot_volumes = oci.pagination.list_call_get_all_results(
                             block_storage.list_boot_volumes,
                             ad['name'],
-                            compartment['id']
+                            compartment['id'],
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                         ).data
 
                     except oci.exceptions.ServiceError as e:
@@ -4318,7 +4337,8 @@ class ShowOCIService(object):
                 try:
                     arrs = oci.pagination.list_call_get_all_results(
                         block_storage.list_volumes, compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -4398,7 +4418,9 @@ class ShowOCIService(object):
                         block_storage.list_volume_groups,
                         compartment['id'],
                         sort_by="DISPLAYNAME",
-                        lifecycle_state=oci.core.models.VolumeGroup.LIFECYCLE_STATE_AVAILABLE).data
+                        lifecycle_state=oci.core.models.VolumeGroup.LIFECYCLE_STATE_AVAILABLE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                    ).data
 
                 except oci.exceptions.ServiceError as e:
                     if self.__check_service_error(e.code):
@@ -4459,7 +4481,8 @@ class ShowOCIService(object):
                         block_storage.list_boot_volume_backups,
                         compartment['id'],
                         sort_by="DISPLAYNAME",
-                        lifecycle_state=oci.core.models.BootVolumeBackup.LIFECYCLE_STATE_AVAILABLE
+                        lifecycle_state=oci.core.models.BootVolumeBackup.LIFECYCLE_STATE_AVAILABLE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -4534,7 +4557,9 @@ class ShowOCIService(object):
                         block_storage.list_volume_backups,
                         compartment['id'],
                         sort_by="DISPLAYNAME",
-                        lifecycle_state=oci.core.models.VolumeBackup.LIFECYCLE_STATE_AVAILABLE).data
+                        lifecycle_state=oci.core.models.VolumeBackup.LIFECYCLE_STATE_AVAILABLE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                    ).data
 
                 except oci.exceptions.ServiceError as e:
                     if self.__check_service_error(e.code):
@@ -5066,7 +5091,9 @@ class ShowOCIService(object):
                     namespace_name = object_storage.get_namespace().data
                     buckets = oci.pagination.list_call_get_all_results(
                         object_storage.list_buckets,
-                        namespace_name, compartment['id']
+                        namespace_name,
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -5210,7 +5237,8 @@ class ShowOCIService(object):
                     stacks = oci.pagination.list_call_get_all_results(
                         orm.list_stacks, compartment_id=compartment['id'],
                         lifecycle_state=oci.resource_manager.models.Stack.LIFECYCLE_STATE_ACTIVE,
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -5233,8 +5261,12 @@ class ShowOCIService(object):
 
                     # check jobs
                     try:
-                        jobs = oci.pagination.list_call_get_all_results(orm.list_jobs, stack_id=stack.id,
-                                                                        sort_by="TIMECREATED").data
+                        jobs = oci.pagination.list_call_get_all_results(
+                            orm.list_jobs,
+                            stack_id=stack.id,
+                            sort_by="TIMECREATED",
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                        ).data
                     except oci.exceptions.ServiceError as e:
                         if self.__check_service_error(e.code):
                             self.__load_print_auth_warning()
@@ -5329,8 +5361,10 @@ class ShowOCIService(object):
                 senders = []
                 try:
                     senders = oci.pagination.list_call_get_all_results(
-                        email.list_senders, compartment['id'],
-                        sort_by="EMAILADDRESS"
+                        email.list_senders,
+                        compartment['id'],
+                        sort_by="EMAILADDRESS",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -5501,7 +5535,8 @@ class ShowOCIService(object):
                             file_storage.list_file_systems,
                             compartment['id'], ad['name'],
                             lifecycle_state=oci.file_storage.models.FileSystemSummary.LIFECYCLE_STATE_ACTIVE,
-                            sort_by="DISPLAYNAME"
+                            sort_by="DISPLAYNAME",
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                         ).data
 
                     except oci.exceptions.ServiceError as e:
@@ -5525,7 +5560,8 @@ class ShowOCIService(object):
                             snapshots = oci.pagination.list_call_get_all_results(
                                 file_storage.list_snapshots,
                                 str(fs.id),
-                                lifecycle_state=oci.file_storage.models.SnapshotSummary.LIFECYCLE_STATE_ACTIVE
+                                lifecycle_state=oci.file_storage.models.SnapshotSummary.LIFECYCLE_STATE_ACTIVE,
+                                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                             ).data
 
                             for snap in snapshots:
@@ -5579,7 +5615,9 @@ class ShowOCIService(object):
                             file_storage.list_mount_targets,
                             compartment['id'], ad['name'],
                             lifecycle_state=oci.file_storage.models.MountTargetSummary.LIFECYCLE_STATE_ACTIVE,
-                            sort_by="DISPLAYNAME").data
+                            sort_by="DISPLAYNAME",
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                        ).data
 
                     except oci.exceptions.ServiceError as e:
                         if self.__check_service_error(e.code):
@@ -5639,7 +5677,8 @@ class ShowOCIService(object):
                         file_storage.list_exports,
                         compartment_id=compartment['id'],
                         lifecycle_state=oci.file_storage.models.ExportSummary.LIFECYCLE_STATE_ACTIVE,
-                        sort_by="PATH"
+                        sort_by="PATH",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -5727,6 +5766,7 @@ class ShowOCIService(object):
 
             # add the key if not exists
             self.__initialize_data_key(self.C_DATABASE, self.C_DATABASE_DBSYSTEMS)
+            self.__initialize_data_key(self.C_DATABASE, self.C_DATABASE_EXADATA)
             self.__initialize_data_key(self.C_DATABASE, self.C_DATABASE_AUTONOMOUS)
             self.__initialize_data_key(self.C_DATABASE, self.C_DATABASE_NOSQL)
             self.__initialize_data_key(self.C_DATABASE, self.C_DATABASE_MYSQL)
@@ -5735,6 +5775,7 @@ class ShowOCIService(object):
             db = self.data[self.C_DATABASE]
 
             # append the data
+            db[self.C_DATABASE_EXADATA] += self.__load_database_exadata_infrastructure(database_client, virtual_network, compartments)
             db[self.C_DATABASE_DBSYSTEMS] += self.__load_database_dbsystems(database_client, virtual_network, compartments)
             db[self.C_DATABASE_AUTONOMOUS] += self.__load_database_autonomouns(database_client, compartments)
             db[self.C_DATABASE_NOSQL] += self.__load_database_nosql(nosql_client, compartments)
@@ -5816,6 +5857,257 @@ class ShowOCIService(object):
             self.__print_error("__load_database_maintatance_windows", e)
 
     ##########################################################################
+    # __load_database_exadata_infrastructure
+    ##########################################################################
+
+    def __load_database_exadata_infrastructure(self, database_client, virtual_network, compartments):
+
+        data = []
+        cnt = 0
+        start_time = time.time()
+
+        try:
+
+            self.__load_print_status("Exadata Infrastructure")
+
+            # loop on all compartments
+            for compartment in compartments:
+                # skip managed paas compartment
+                if self.__if_managed_paas_compartment(compartment['name']):
+                    print(".", end="")
+                    continue
+
+                print(".", end="")
+
+                # list db system
+                list_exa = []
+                try:
+                    list_exa = oci.pagination.list_call_get_all_results(
+                        database_client.list_cloud_exadata_infrastructures,
+                        compartment['id'],
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                    ).data
+
+                except oci.exceptions.ServiceError as e:
+                    if self.__check_service_error(e.code):
+                        self.__load_print_auth_warning()
+                        continue
+                    else:
+                        raise
+
+                # loop on the Exadata infrastructure
+                # dbs = oci.database.models.CloudExadataInfrastructureSummary
+                for dbs in list_exa:
+                    if (dbs.lifecycle_state == oci.database.models.CloudExadataInfrastructureSummary.LIFECYCLE_STATE_TERMINATED or
+                            dbs.lifecycle_state == oci.database.models.CloudExadataInfrastructureSummary.LIFECYCLE_STATE_TERMINATING):
+                        continue
+
+                    value = {'id': str(dbs.id),
+                             'display_name': str(dbs.display_name),
+                             'shape': str(dbs.shape),
+                             'shape_ocpu': 0,
+                             'shape_memory_gb': 0,
+                             'shape_storage_tb': 0,
+                             'version': 'XP',
+                             'lifecycle_state': str(dbs.lifecycle_state),
+                             'lifecycle_details': str(dbs.lifecycle_details),
+                             'availability_domain': str(dbs.availability_domain),
+                             'compute_count': str(dbs.compute_count),
+                             'storage_count': str(dbs.storage_count),
+                             'total_storage_size_in_gbs': str(dbs.total_storage_size_in_gbs),
+                             'available_storage_size_in_gbs': str(dbs.available_storage_size_in_gbs),
+                             'compartment_name': str(compartment['name']),
+                             'compartment_id': str(compartment['id']),
+                             'time_created': str(dbs.time_created),
+                             'last_maintenance_run': self.__load_database_maintatance(database_client, dbs.last_maintenance_run_id, str(dbs.display_name) + " - " + str(dbs.shape)),
+                             'next_maintenance_run': self.__load_database_maintatance(database_client, dbs.next_maintenance_run_id, str(dbs.display_name) + " - " + str(dbs.shape)),
+                             'maintenance_window': self.__load_database_maintatance_windows(dbs.maintenance_window),
+                             'defined_tags': [] if dbs.defined_tags is None else dbs.defined_tags,
+                             'freeform_tags': [] if dbs.freeform_tags is None else dbs.freeform_tags,
+                             'region_name': str(self.config['region']),
+                             'vm_clusters': self.__load_database_exadata_vm_clusters(database_client, virtual_network, dbs.id, compartment)
+                             }
+
+                    # get shape
+                    if dbs.shape:
+                        shape_sizes = self.get_shape_details(str(dbs.shape))
+                        if shape_sizes:
+                            value['shape_ocpu'] = shape_sizes['cpu']
+                            value['shape_memory_gb'] = shape_sizes['memory']
+                            value['shape_storage_tb'] = shape_sizes['storage']
+
+                        # if x8m calculate ocpu and storage
+                        if dbs.shape == "Exadata.X8M":
+                            if dbs.compute_count != "2" or dbs.storage_count != "3":
+                                value['shape_ocpu'] = dbs.compute_count * 50
+                                value['shape_storage_tb'] = dbs.storage_count * 49.5
+                                value['shape_memory_gb'] = dbs.compute_count * 720
+
+                    # add the data
+                    cnt += 1
+                    data.append(value)
+
+            self.__load_print_cnt(cnt, start_time)
+            return data
+
+        except oci.exceptions.RequestException as e:
+            if self.__check_request_error(e):
+                return data
+            raise
+        except Exception as e:
+            self.__print_error("__load_database_exadata_infrastructure", e)
+            return data
+
+    ##########################################################################
+    # __load_database_exadata_vm_clusters
+    ##########################################################################
+    def __load_database_exadata_vm_clusters(self, database_client, virtual_network, exa_id, compartment):
+
+        data = []
+        try:
+            vms = database_client.list_cloud_vm_clusters(
+                compartment['id'],
+                cloud_exadata_infrastructure_id=exa_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+            ).data
+
+            # arr = oci.database.models.CloudVmClusterSummary
+            for arr in vms:
+                if (arr.lifecycle_state == oci.database.models.CloudVmClusterSummary.LIFECYCLE_STATE_TERMINATED or
+                        arr.lifecycle_state == oci.database.models.CloudVmClusterSummary.LIFECYCLE_STATE_TERMINATING):
+                    continue
+
+                value = {
+                    'id': str(arr.id),
+                    'cluster_name': str(arr.cluster_name),
+                    'hostname': str(arr.hostname),
+                    'compartment_id': str(arr.compartment_id),
+                    'availability_domain': str(arr.availability_domain),
+                    'data_subnet_id': str(arr.subnet_id),
+                    'data_subnet': self.get_network_subnet(str(arr.subnet_id), True),
+                    'backup_subnet_id': str(arr.backup_subnet_id),
+                    'backup_subnet': "" if arr.backup_subnet_id is None else self.get_network_subnet(str(arr.backup_subnet_id), True),
+                    'nsg_ids': str(arr.nsg_ids),
+                    'backup_network_nsg_ids': str(arr.backup_network_nsg_ids),
+                    'last_update_history_entry_id': str(arr.last_update_history_entry_id),
+                    'shape': str(arr.shape),
+                    'listener_port': str(arr.listener_port),
+                    'lifecycle_state': str(arr.lifecycle_state),
+                    'node_count': str(arr.node_count),
+                    'storage_size_in_gbs': str(arr.storage_size_in_gbs),
+                    'display_name': str(arr.display_name),
+                    'time_created': str(arr.time_created),
+                    'lifecycle_details': str(arr.lifecycle_details),
+                    'time_zone': str(arr.time_zone),
+                    'domain': str(arr.domain),
+                    'cpu_core_count': str(arr.cpu_core_count),
+                    'data_storage_percentage': str(arr.data_storage_percentage),
+                    'is_local_backup_enabled': str(arr.is_local_backup_enabled),
+                    'is_sparse_diskgroup_enabled': str(arr.is_sparse_diskgroup_enabled),
+                    'gi_version': str(arr.gi_version),
+                    'system_version': str(arr.system_version),
+                    'ssh_public_keys': str(arr.ssh_public_keys),
+                    'license_model': str(arr.license_model),
+                    'disk_redundancy': str(arr.disk_redundancy),
+                    'scan_ip_ids': str(arr.scan_ip_ids),
+                    'vip_ids': str(arr.vip_ids),
+                    'scan_dns_record_id': str(arr.scan_dns_record_id),
+                    'defined_tags': [] if arr.defined_tags is None else arr.defined_tags,
+                    'freeform_tags': [] if arr.freeform_tags is None else arr.freeform_tags,
+                    'patches': [],
+                    'db_homes': self.__load_database_dbsystems_dbhomes(database_client, virtual_network, compartment, arr.id, exa=True),
+                    'db_nodes': self.__load_database_dbsystems_dbnodes(database_client, virtual_network, compartment, arr.id, exa=True),
+                    'region_name': str(self.config['region']),
+                    'scan_ips': [],
+                    'vip_ips': []
+                }
+
+                # Skip the patches, there is an issue with the api for the vm cluster
+                # value['patches'] = self.__load_database_exadata_vm_patches(database_client, arr.id),
+
+                # get shape
+                if arr.shape:
+                    shape_sizes = self.get_shape_details(str(arr.shape))
+                    if shape_sizes:
+                        value['shape_ocpu'] = shape_sizes['cpu']
+                        value['shape_memory_gb'] = shape_sizes['memory']
+                        value['shape_storage_tb'] = shape_sizes['storage']
+
+                # license model
+                if arr.license_model == oci.database.models.CloudVmClusterSummary.LICENSE_MODEL_LICENSE_INCLUDED:
+                    value['license_model'] = "INCL"
+                elif arr.license_model == oci.database.models.CloudVmClusterSummary.LICENSE_MODEL_BRING_YOUR_OWN_LICENSE:
+                    value['license_model'] = "BYOL"
+                else:
+                    value['license_model'] = str(arr.license_model)
+
+                # scan IPs
+                if arr.scan_ip_ids is not None:
+                    scan_ips = []
+                    for scan_ip in arr.scan_ip_ids:
+                        scan_ips.append(self.__load_core_network_single_privateip(virtual_network, scan_ip))
+                    value['scan_ips'] = scan_ips
+
+                # VIPs
+                if arr.vip_ids is not None:
+                    vip_ips = []
+                    for vipip in arr.vip_ids:
+                        vip_ips.append(self.__load_core_network_single_privateip(virtual_network, vipip))
+                    value['vip_ips'] = vip_ips
+
+                # add to main data
+                data.append(value)
+
+            return data
+
+        except oci.exceptions.ServiceError as e:
+            if self.__check_service_error(e.code):
+                self.__load_print_auth_warning()
+                return data
+            else:
+                raise
+        except oci.exceptions.RequestException as e:
+            if self.__check_request_error(e):
+                return data
+            raise
+        except Exception as e:
+            self.__print_error("__load_database_exadata_vm_clusters", e)
+            return data
+
+    ##########################################################################
+    # __load_database_exadata_vm_patches
+    ##########################################################################
+    def __load_database_exadata_vm_patches(self, database_client, vm_id):
+
+        data = []
+        try:
+            dbps = oci.pagination.list_call_get_all_results(
+                database_client.list_vm_cluster_patches,
+                vm_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+            ).data
+
+            for dbp in dbps:
+                data.append({'id': dbp.id, 'description': str(dbp.description),
+                             'version': str(dbp.version), 'time_released': str(dbp.time_released),
+                             'last_action': str(dbp.last_action)})
+            return data
+
+        except oci.exceptions.ServiceError as e:
+            if self.__check_service_error(e.code):
+                return data
+            else:
+                raise
+        except oci.exceptions.RequestException as e:
+            if self.__check_request_error(e):
+                return data
+            raise
+        except Exception as e:
+            self.__print_error("__load_database_exadata_vm_patches", e)
+            return data
+
+    ##########################################################################
     # __load_database_dbsystems
     ##########################################################################
 
@@ -5844,7 +6136,8 @@ class ShowOCIService(object):
                     list_db_systems = oci.pagination.list_call_get_all_results(
                         database_client.list_db_systems,
                         compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -5857,7 +6150,7 @@ class ShowOCIService(object):
                 # loop on the db systems
                 # dbs = oci.database.models.DbSystemSummary
                 for dbs in list_db_systems:
-                    if dbs.lifecycle_state == oci.database.models.DbSystemSummary.LIFECYCLE_STATE_TERMINATED:
+                    if (dbs.lifecycle_state == oci.database.models.DbSystemSummary.LIFECYCLE_STATE_TERMINATED or dbs.lifecycle_state == "MIGRATED"):
                         continue
 
                     value = {'id': str(dbs.id),
@@ -5896,8 +6189,9 @@ class ShowOCIService(object):
                              'defined_tags': [] if dbs.defined_tags is None else dbs.defined_tags,
                              'freeform_tags': [] if dbs.freeform_tags is None else dbs.freeform_tags,
                              'patches': self.__load_database_dbsystems_patches(database_client, dbs.id),
-                             'db_nodes': self.__load_database_dbsystems_dbnodes(database_client, virtual_network, dbs.id, compartment),
-                             'db_homes': self.__load_database_dbsystems_dbhomes(database_client, virtual_network, dbs.id, compartment)}
+                             'db_nodes': self.__load_database_dbsystems_dbnodes(database_client, virtual_network, compartment, dbs.id),
+                             'db_homes': self.__load_database_dbsystems_dbhomes(database_client, virtual_network, compartment, dbs.id)
+                             }
 
                     # get shape
                     if dbs.shape:
@@ -5964,13 +6258,28 @@ class ShowOCIService(object):
             return data
 
     ##########################################################################
-    # __load_database_dbsystems_dbnodes
+    # __load_database_exadata_infrastructure
     ##########################################################################
-    def __load_database_dbsystems_dbnodes(self, database_client, virtual_network, dbs_id, compartment):
+    def __load_database_dbsystems_dbnodes(self, database_client, virtual_network, compartment, dbs_id, exa=False):
 
         data = []
+        db_nodes = []
+        api_call = ""
         try:
-            db_nodes = database_client.list_db_nodes(compartment['id'], db_system_id=dbs_id).data
+            if not exa:
+                api_call = "database_client.list_db_nodes with db_system_id"
+                db_nodes = database_client.list_db_nodes(
+                    compartment['id'],
+                    db_system_id=dbs_id,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                ).data
+            else:
+                api_call = "database_client.list_db_nodes with vm_cluster_id"
+                db_nodes = database_client.list_db_nodes(
+                    compartment['id'],
+                    vm_cluster_id=dbs_id,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                ).data
 
             # db_node = oci.database.models.DbNodeSummary
             for db_node in db_nodes:
@@ -6000,23 +6309,43 @@ class ShowOCIService(object):
                 self.__load_print_auth_warning()
                 return data
             else:
+                print("Error at API " + api_call)
                 raise
         except oci.exceptions.RequestException as e:
             if self.__check_request_error(e):
                 return data
-            raise
+            else:
+                print("Error at API " + api_call)
+                raise
         except Exception as e:
-            self.__print_error("__load_database_dbsystems_dbnodes", e)
+            self.__print_error("__load_database_dbsystems_dbnodes, API=" + api_call, e)
             return data
 
     ##########################################################################
     # __load_database_dbsystems_dbhomes
     ##########################################################################
-    def __load_database_dbsystems_dbhomes(self, database_client, virtual_network, dbs_id, compartment):
+    def __load_database_dbsystems_dbhomes(self, database_client, virtual_network, compartment, dbs_id, exa=False):
 
         data = []
+        db_homes = []
+        api_call = ""
         try:
-            db_homes = oci.pagination.list_call_get_all_results(database_client.list_db_homes, compartment['id'], db_system_id=dbs_id).data
+            if not exa:
+                api_call = "database_client.list_db_homes with db_system_id"
+                db_homes = oci.pagination.list_call_get_all_results(
+                    database_client.list_db_homes,
+                    compartment['id'],
+                    db_system_id=dbs_id,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                ).data
+            else:
+                api_call = "database_client.list_db_homes with vm_cluster_id"
+                db_homes = oci.pagination.list_call_get_all_results(
+                    database_client.list_db_homes,
+                    compartment['id'],
+                    vm_cluster_id=dbs_id,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                ).data
 
             # db_home = oci.database.models.DbHomeSummary
             for db_home in db_homes:
@@ -6027,6 +6356,7 @@ class ShowOCIService(object):
                      'last_patch_history_entry_id': str(db_home.last_patch_history_entry_id),
                      'lifecycle_state': str(db_home.lifecycle_state),
                      'db_system_id': str(db_home.db_system_id),
+                     'vm_cluster_id': str(db_home.vm_cluster_id),
                      'db_version': str(db_home.db_version),
                      'time_created': str(db_home.time_created),
                      'databases': self.__load_database_dbsystems_dbhomes_databases(database_client, db_home.id, compartment),
@@ -6040,13 +6370,16 @@ class ShowOCIService(object):
                 self.__load_print_auth_warning("h")
                 return data
             else:
+                print("Error at API " + api_call)
                 raise
         except oci.exceptions.RequestException as e:
             if self.__check_request_error(e):
                 return data
-            raise
+            else:
+                print("Error at API " + api_call)
+                raise
         except Exception as e:
-            self.__print_error("__load_database_dbsystems_dbhomess", e)
+            self.__print_error("__load_database_dbsystems_dbhomess, API=" + api_call, e)
             return data
 
     ##########################################################################
@@ -6061,7 +6394,8 @@ class ShowOCIService(object):
                 database_client.list_databases,
                 compartment['id'],
                 db_home_id=db_home_id,
-                sort_by="DBNAME"
+                sort_by="DBNAME",
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
             ).data
 
             # db = oci.database.models.DatabaseSummary
@@ -6121,7 +6455,8 @@ class ShowOCIService(object):
         try:
             dbps = oci.pagination.list_call_get_all_results(
                 database_client.list_db_home_patches,
-                dbhome_id
+                dbhome_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
             ).data
 
             for dbp in dbps:
@@ -6153,7 +6488,11 @@ class ShowOCIService(object):
 
         data = []
         try:
-            dbps = oci.pagination.list_call_get_all_results(database_client.list_db_system_patches, dbs_id).data
+            dbps = oci.pagination.list_call_get_all_results(
+                database_client.list_db_system_patches,
+                dbs_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+            ).data
 
             for dbp in dbps:
                 data.append({'id': dbp.id, 'description': str(dbp.description),
@@ -6181,7 +6520,11 @@ class ShowOCIService(object):
 
         data = []
         try:
-            backups = oci.pagination.list_call_get_all_results(database_client.list_backups, database_id=db_id).data
+            backups = oci.pagination.list_call_get_all_results(
+                database_client.list_backups,
+                database_id=db_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+            ).data
 
             for backup in backups:
                 data.append(
@@ -6216,7 +6559,11 @@ class ShowOCIService(object):
 
         data = []
         try:
-            dgs = oci.pagination.list_call_get_all_results(database_client.list_data_guard_associations, database_id=db_id).data
+            dgs = oci.pagination.list_call_get_all_results(
+                database_client.list_data_guard_associations,
+                database_id=db_id,
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+            ).data
 
             # dg = oci.database.models.DataGuardAssociationSummary
             for dg in dgs:
@@ -6294,7 +6641,8 @@ class ShowOCIService(object):
                     list_autos = oci.pagination.list_call_get_all_results(
                         database_client.list_autonomous_databases,
                         compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -6496,7 +6844,8 @@ class ShowOCIService(object):
                     databases = oci.pagination.list_call_get_all_results(
                         mysql_client.list_db_systems,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 # mysql throw service error often, ignoring incase it does
@@ -6597,7 +6946,8 @@ class ShowOCIService(object):
             backups = oci.pagination.list_call_get_all_results(
                 database_client.list_autonomous_database_backups,
                 autonomous_database_id=db_id,
-                sort_by="TIMECREATED"
+                sort_by="TIMECREATED",
+                retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
             ).data
 
             # backup = oci.database.models.AutonomousDatabaseBackupSummary
@@ -6690,8 +7040,10 @@ class ShowOCIService(object):
 
                 try:
                     list_clusters = oci.pagination.list_call_get_all_results(
-                        container_client.list_node_pools, compartment['id'],
-                        sort_by="NAME"
+                        container_client.list_node_pools,
+                        compartment['id'],
+                        sort_by="NAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -6753,8 +7105,10 @@ class ShowOCIService(object):
 
                 try:
                     list_clusters = oci.pagination.list_call_get_all_results(
-                        container_client.list_clusters, compartment['id'],
-                        sort_by="NAME"
+                        container_client.list_clusters,
+                        compartment['id'],
+                        sort_by="NAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -6854,7 +7208,8 @@ class ShowOCIService(object):
                     streams = oci.pagination.list_call_get_all_results(
                         stream_client.list_streams, compartment_id=compartment['id'],
                         sort_by="NAME",
-                        lifecycle_state=oci.streaming.models.StreamSummary.LIFECYCLE_STATE_ACTIVE
+                        lifecycle_state=oci.streaming.models.StreamSummary.LIFECYCLE_STATE_ACTIVE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -6954,7 +7309,8 @@ class ShowOCIService(object):
                 try:
                     apigs = oci.pagination.list_call_get_all_results(
                         api_client.list_gateways, compartment_id=compartment['id'],
-                        lifecycle_state="ACTIVE"
+                        lifecycle_state="ACTIVE",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7062,7 +7418,8 @@ class ShowOCIService(object):
                     apps = oci.pagination.list_call_get_all_results(
                         function_client.list_applications, compartment_id=compartment['id'],
                         sort_by="displayName",
-                        lifecycle_state='ACTIVE'
+                        lifecycle_state='ACTIVE',
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7157,7 +7514,8 @@ class ShowOCIService(object):
                 budgets = oci.pagination.list_call_get_all_results(
                     budget_client.list_budgets,
                     tenancy_id,
-                    lifecycle_state=oci.budget.models.BudgetSummary.LIFECYCLE_STATE_ACTIVE
+                    lifecycle_state=oci.budget.models.BudgetSummary.LIFECYCLE_STATE_ACTIVE,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                 ).data
 
             except oci.exceptions.ServiceError as e:
@@ -7300,7 +7658,8 @@ class ShowOCIService(object):
                         event_client.list_rules,
                         compartment['id'],
                         sort_by="DISPLAY_NAME",
-                        lifecycle_state="ACTIVE"
+                        lifecycle_state="ACTIVE",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7368,7 +7727,8 @@ class ShowOCIService(object):
                         monitor_client.list_alarms,
                         compartment['id'],
                         sort_by="displayName",
-                        lifecycle_state=oci.monitoring.models.Alarm.LIFECYCLE_STATE_ACTIVE
+                        lifecycle_state=oci.monitoring.models.Alarm.LIFECYCLE_STATE_ACTIVE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7438,7 +7798,8 @@ class ShowOCIService(object):
                         ons_cp_client.list_topics,
                         compartment['id'],
                         sort_by="TIMECREATED",
-                        lifecycle_state=oci.ons.models.NotificationTopicSummary.LIFECYCLE_STATE_ACTIVE
+                        lifecycle_state=oci.ons.models.NotificationTopicSummary.LIFECYCLE_STATE_ACTIVE,
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7504,7 +7865,8 @@ class ShowOCIService(object):
                 try:
                     subs = oci.pagination.list_call_get_all_results(
                         ons_dp_client.list_subscriptions,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7633,7 +7995,8 @@ class ShowOCIService(object):
                     healthchecks = oci.pagination.list_call_get_all_results(
                         healthcheck_client.list_ping_monitors,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7713,7 +8076,8 @@ class ShowOCIService(object):
                     healthchecks = oci.pagination.list_call_get_all_results(
                         healthcheck_client.list_http_monitors,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7801,7 +8165,8 @@ class ShowOCIService(object):
                         dns_client.list_zones,
                         compartment['id'],
                         lifecycle_state=oci.dns.models.ZoneSummary.LIFECYCLE_STATE_ACTIVE,
-                        sort_by="name"
+                        sort_by="name",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7872,7 +8237,8 @@ class ShowOCIService(object):
                         dns_client.list_steering_policies,
                         compartment['id'],
                         lifecycle_state=oci.dns.models.SteeringPolicySummary.LIFECYCLE_STATE_ACTIVE,
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -7941,7 +8307,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         waas.list_waas_policies,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8069,7 +8436,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         dc_client.list_catalogs,
                         compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8143,7 +8511,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         ds_client.list_projects,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8216,7 +8585,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         df_client.list_applications,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8290,7 +8660,8 @@ class ShowOCIService(object):
                 try:
                     odas = oci.pagination.list_call_get_all_results(
                         oda_client.list_oda_instances,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8363,7 +8734,8 @@ class ShowOCIService(object):
                 try:
                     bdss = oci.pagination.list_call_get_all_results(
                         bds_client.list_bds_instances,
-                        compartment['id']
+                        compartment['id'],
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 # TBD: don't add warning count until GA on the service
@@ -8490,7 +8862,8 @@ class ShowOCIService(object):
                     oics = oci.pagination.list_call_get_all_results(
                         oic_client.list_integration_instances,
                         compartment['id'],
-                        sort_by="DISPLAYNAME"
+                        sort_by="DISPLAYNAME",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8564,7 +8937,8 @@ class ShowOCIService(object):
                     oacs = oci.pagination.list_call_get_all_results(
                         oac_client.list_analytics_instances,
                         compartment['id'],
-                        sort_by="name"
+                        sort_by="name",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8640,7 +9014,8 @@ class ShowOCIService(object):
                     oces = oci.pagination.list_call_get_all_results(
                         oce_client.list_oce_instances,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -8873,7 +9248,13 @@ class ShowOCIService(object):
                     # get the limits per service
                     limits = []
                     try:
-                        limits = oci.pagination.list_call_get_all_results(limits_client.list_limit_values, tenancy_id, service_name=service.name, sort_by="name").data
+                        limits = oci.pagination.list_call_get_all_results(
+                            limits_client.list_limit_values,
+                            tenancy_id,
+                            service_name=service.name,
+                            sort_by="name",
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
+                        ).data
                     except oci.exceptions.Exception as e:
                         if self.__check_service_error(e.code):
                             self.__load_print_auth_warning("a", False)
@@ -9088,7 +9469,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         cg_client.list_targets,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -9164,7 +9546,8 @@ class ShowOCIService(object):
                     array = oci.pagination.list_call_get_all_results(
                         log_client.list_log_groups,
                         compartment['id'],
-                        sort_by="displayName"
+                        sort_by="displayName",
+                        retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                     ).data
 
                 except oci.exceptions.ServiceError as e:
@@ -9204,7 +9587,8 @@ class ShowOCIService(object):
                         logs = oci.pagination.list_call_get_all_results(
                             log_client.list_logs,
                             item.id,
-                            sort_by="displayName"
+                            sort_by="displayName",
+                            retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                         ).data
 
                     except oci.exceptions.ServiceError as e: